requirements.txt
future
numpy
pandas
pyyaml
six
sklearn
tensorboard
torch
torchvision
